requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Werkzeug==3.0.1
psycopg[binary]==3.2.10
python-dotenv==1.0.0